cricket-analyzer/requirements.txt
# core
opencv-python
mediapipe
numpy
pyyaml
aiofiles
